refactor: extract common ui test logic into macro
