+++ crates/internal_test_util/Cargo.toml
@@ -9,3 +9,5 @@ publish = false
 [dependencies]
 bevy_app = { workspace = true }
 bevy_internal = { workspace = true }
+internal_test_proc_macro = { workspace = true }
+trybuild = { workspace = true }
